Add validation and error handling

+++ src/controllers/movieController.js
@@ -2,6 +2,7 @@ import { Router } from "express";
 import movieService from "../services/movieService.js";
 import castService from "../services/castService.js";
 import { isAuth } from "../middlewares/authMiddleware.js";
+import { getErrorMessage } from "../utils/errorUtils.js";
 
 const router = Router();
 
@@ -17,7 +18,14 @@ router.post('/create', isAuth, async (req, res) => {
     const movieData = req.body;
     const ownerId = req.user?._id;
 
-    await movieService.create(movieData, ownerId);
+    try {
+        await movieService.create(movieData, ownerId);
+        
+    } catch (err) {
+        const errorMessage = getErrorMessage(err);
+        return res.render('movies/create', { error: errorMessage, movie: movieData });
+        
+    }
 
     res.redirect('/');
 })
@@ -57,6 +65,14 @@ router.post('/:movieId/attach', isAuth, async (req, res) => {
 
 router.get('/:movieId/delete', isAuth, async (req, res) => {
     const movieId = req.params.movieId;
+
+    // check if owner
+    const movie = await movieService.getOne(movieId).lean();
+    if (movie.owner?.toString() !== req.user._id) {
+        // return res.render('movies/details', { movie, isOwner: false, error: 'You cannot delete this movie!' });
+        res.setError('You cannot delete this movie!');
+        return res.redirect('/404');
+    }
 
     await movieService.remove(movieId);
